+++ crates/viewer/re_space_view_graph/src/layout/mod.rs
@@ -1,7 +1,7 @@
-mod layout;
 mod provider;
 mod request;
+mod result;
 
-pub use layout::Layout;
 pub use provider::ForceLayoutProvider;
 pub use request::LayoutRequest;
+pub use result::Layout;
