--- crates/viewer/re_space_view_graph/src/ui/draw/mod.rs
@@ -1,3 +1,0 @@
-mod node;
-
-pub use node::{CircleLabel, TextLabel, DrawableLabel};
