+++ crates/viewer/re_space_view_graph/src/ui/mod.rs
@@ -1,10 +1,5 @@
-pub mod draw;
+mod draw;
 mod state;
 
-pub mod scene;
-
+pub use draw::{draw_debug, draw_edge, draw_entity_rect, draw_node, DrawableLabel};
 pub use state::GraphSpaceViewState;
-
-pub fn bounding_rect_from_iter(rectangles: impl Iterator<Item = egui::Rect>) -> egui::Rect {
-    rectangles.fold(egui::Rect::NOTHING, |acc, rect| acc.union(rect))
-}
