+++ crates/viewer/re_ui/src/lib.rs
@@ -7,7 +7,6 @@ mod syntax_highlighting;
 
 mod context_ext;
 pub mod drag_and_drop;
-pub mod zoom_pan_area;
 pub mod icons;
 pub mod list_item;
 mod markdown_utils;
@@ -15,6 +14,7 @@ pub mod modal;
 mod section_collapsing_header;
 pub mod toasts;
 mod ui_ext;
+pub mod zoom_pan_area;
 
 use egui::NumExt as _;
